Moved the ResourceOperation to the Operation folder to resolve recurrsive import errors.
