+++ pennylane/resource/__init__.py
@@ -70,7 +70,6 @@ Resource Classes
     :toctree: api
 
     ~Resources
-    ~ResourcesOperation
 
 Tracking Resources for Custom Operations
 ----------------------------------------
@@ -122,6 +121,6 @@ gate_sizes:
 from .error import AlgorithmicError, ErrorOperation, SpectralNormError
 from .first_quantization import FirstQuantization
 from .measurement import estimate_error, estimate_shots
-from .resource import Resources, ResourcesOperation
+from .resource import Resources
 from .second_quantization import DoubleFactorization
 from .specs import specs
