+++ pennylane/resource/resource.py
@@ -19,7 +19,7 @@ from collections import defaultdict
 from dataclasses import dataclass, field
 
 from pennylane.measurements import Shots
-from pennylane.operation import Operation
+from pennylane.operation import ResourcesOperation
 
 
 @dataclass(frozen=True)
@@ -87,42 +87,6 @@ class Resources:
     def _ipython_display_(self):
         """Displays __str__ in ipython instead of __repr__"""
         print(str(self))
-
-
-class ResourcesOperation(Operation):
-    r"""Base class that represents quantum gates or channels applied to quantum
-    states and stores the resource requirements of the quantum gate.
-
-    .. note::
-        Child classes must implement the :func:`~.ResourcesOperation.resources` method which computes
-        the resource requirements of the operation.
-    """
-
-    @abstractmethod
-    def resources(self) -> Resources:
-        r"""Compute the resources required for this operation.
-
-        Returns:
-            Resources: The resources required by this operation.
-
-        **Examples**
-
-        >>> class CustomOp(ResourcesOperation):
-        ...     num_wires = 2
-        ...     def resources(self):
-        ...         return Resources(num_wires=self.num_wires, num_gates=3, depth=2)
-        ...
-        >>> op = CustomOp(wires=[0, 1])
-        >>> print(op.resources())
-        wires: 2
-        gates: 3
-        depth: 2
-        shots: Shots(total=None)
-        gate_types:
-        {}
-        gate_sizes:
-        {}
-        """
 
 
 def _count_resources(tape) -> Resources:
